:recycle: Moved zustand store defs to own files
