+++ package.json
@@ -39,7 +39,10 @@
     "redux-thunk": "^2.3.0",
     "resize-observer-polyfill": "^1.5.1",
     "uuid": "^8.3.2",
-    "zustand": "^3.5.5"
+    "y-webrtc": "^10.2.0",
+    "yjs": "^13.5.11",
+    "zustand": "^3.5.5",
+    "zustand-middleware-yjs": "^1.2.1"
   },
   "devDependencies": {
     "@commitlint/cli": "^11.0.0",
